src/main/java/io/surisoft/capi/schema/MappingId.java
package io.surisoft.capi.schema;

import java.io.Serializable;

public class MappingId implements Serializable {
    private String rootContext;
    private String hostname;
    private int port;

    public String getRootContext() {
        return rootContext;
    }

    public void setRootContext(String rootContext) {
        this.rootContext = rootContext;
    }

    public String getHostname() {
        return hostname;
    }

    public void setHostname(String hostname) {
        this.hostname = hostname;
    }

    public int getPort() {
        return port;
    }

    public void setPort(int port) {
        this.port = port;
    }
}
